move make sweep
