+++ qcodes/parameters/sweep_values.py
@@ -1,7 +1,19 @@
 from copy import deepcopy
-from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Sequence, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Dict,
+    Iterator,
+    List,
+    Optional,
+    Sequence,
+    Union,
+    cast,
+)
 
-from qcodes.utils.helpers import make_sweep, named_repr
+import numpy as np
+
+from qcodes.utils.helpers import named_repr
 from qcodes.utils.metadata import Metadatable
 
 from .permissive_range import permissive_range
@@ -9,6 +21,67 @@ from .sequence_helpers import is_sequence
 
 if TYPE_CHECKING:
     from qcodes.parameters import ParameterBase
+
+
+# This is very much related to the permissive_range but more
+# strict on the input, start and endpoints are always included,
+# and a sweep is only created if the step matches an integer
+# number of points.
+# numpy is a dependency anyways.
+# Furthermore the sweep allows to take a number of points and generates
+# an array with endpoints included, which is more intuitive to use in a sweep.
+def make_sweep(
+    start: float, stop: float, step: Optional[float] = None, num: Optional[int] = None
+) -> List[float]:
+    """
+    Generate numbers over a specified interval.
+    Requires ``start`` and ``stop`` and (``step`` or ``num``).
+    The sign of ``step`` is not relevant.
+
+    Args:
+        start: The starting value of the sequence.
+        stop: The end value of the sequence.
+        step:  Spacing between values.
+        num: Number of values to generate.
+
+    Returns:
+        numpy.ndarray: numbers over a specified interval as a ``numpy.linspace``.
+
+    Examples:
+        >>> make_sweep(0, 10, num=5)
+        [0.0, 2.5, 5.0, 7.5, 10.0]
+        >>> make_sweep(5, 10, step=1)
+        [5.0, 6.0, 7.0, 8.0, 9.0, 10.0]
+        >>> make_sweep(15, 10.5, step=1.5)
+        >[15.0, 13.5, 12.0, 10.5]
+    """
+    if step and num:
+        raise AttributeError("Don't use `step` and `num` at the same time.")
+    if (step is None) and (num is None):
+        raise ValueError(
+            "If you really want to go from `start` to "
+            "`stop` in one step, specify `num=2`."
+        )
+    if step is not None:
+        steps = abs((stop - start) / step)
+        tolerance = 1e-10
+        steps_lo = int(np.floor(steps + tolerance))
+        steps_hi = int(np.ceil(steps - tolerance))
+
+        if steps_lo != steps_hi:
+            raise ValueError(
+                "Could not find an integer number of points for "
+                "the the given `start`, `stop`, and `step` "
+                "values. \nNumber of points is {:d} or {:d}.".format(
+                    steps_lo + 1, steps_hi + 1
+                )
+            )
+        num_steps = steps_lo + 1
+    elif num is not None:
+        num_steps = num
+
+    output_list = np.linspace(start, stop, num=num_steps).tolist()
+    return cast(List[float], output_list)
 
 
 class SweepValues(Metadatable):
